fix(sdei): fix coverity finding  array index read
Fix - CID 463142: (#1 of 1): Overflowed array index
read (INTEGER_OVERFLOW)

Based on issue -
https://scan4.scan.coverity.com/#/project-view/68818/11439?selectedIssue=463142

Add boundary checks overflow checks.

+++ services/std_svc/sdei/sdei_main.c
@@ -947,7 +947,7 @@ static int sdei_signal(int ev_num, uint64_t target_pe)
 		for (i = 0; i < SDEI_PRIVATE_MAPPING()->num_maps; i++) {
 			if (map_priv->ev_num == SDEI_EVENT_0) {
 				se = get_event_entry_target_pe((long int) i,
-				(unsigned int) SDEI_PRIVATE_MAPPING()->num_maps, target_pe);
+					SDEI_PRIVATE_MAPPING()->num_maps, target_pe);
 				if (se == NULL) {
 					return SDEI_EINVAL;
 				}
